+++ js/const.js
@@ -1,0 +1,20 @@
+/*eslint-disable no-unused-vars */
+export const PICTURE_COUNT = 25;
+export const AVATAR_COUNT = 6;
+export const LIKE_MIN_COUNT = 15 ;
+export const LIKE_MAX_COUNT = 200;
+export const COMMENT_LINES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+export const DESCRIPTIONS = ['проводим классные выходные',
+  'тестим новую камеру',
+  'каникулы улет',
+  'фоткаемся',
+  'клааааааас',
+  'жгет супер лето'];
+export const NAMES = ['вован', 'максим','яникс', 'сергей','кристина','маша'];
